a bit more refactoring of homescreen
