+++ lib/features/ar_and_products/presentation/screens/category_products_screen.dart
