+++ lib/features/ideabook/presentation/widgets/my_ideas_tab.dart
@@ -1,6 +1,7 @@
 import 'package:decor_ride/app/theme_extension.dart';
 import 'package:decor_ride/features/ideabook/data/data_sources/general_ideas_datasource_impl.dart';
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class MyIdeasTab extends ConsumerWidget {
@@ -21,7 +22,9 @@ class MyIdeasTab extends ConsumerWidget {
               mainAxisAlignment: MainAxisAlignment.end,
               children: [
                 IconButton(
-                  onPressed: () {},
+                  onPressed: () {
+                    context.push('/ar_main_view');
+                  },
                   icon: const Icon(
                     Icons.camera_alt_rounded,
                     size: 20.0,
